// sla-app/src/app/models/DB-Data.model.ts
export class DBDatas{

  SLAId: number;
  SLATierId: number;
  SLANumber: string;
  SLADefinition: string;
  SLATier: string;
  M1SLAPercent: number;
  M2SLAPercent: number;
  M3SLAPercent: number;
  M4SLAPercent: number;
  M5SLAPercent: number;
  M6SLAPercent: number;
  M7SLAPercent: number;
  M8SLAPercent: number;
  M9SLAPercent: number;
  M10SLAPercent: number;
  M11SLAPercent: number;
  M12SLAPercent: number;
TargetPercentage: number;  
}